Extract the refresh_callback() and token_callback() views
Extract the refresh_callback() and token_callback() Pyramid views into a
new views/oauth.py file.

The oauth_callback() helper function that they both call must also be
extracted.
