+++ lti/views/__init__.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+
+
+def includeme(config):
+    config.scan(__name__)
